backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

streamlit==1.36.0
requests==2.32.3

langchain==0.2.7
langchain-community==0.2.7
langchain-openai==0.1.14

openai==1.35.13

python-dotenv==1.0.1

pdfplumber==0.11.4
python-docx==1.1.2

pydantic==2.7.4
